Adds a lot of juicyness to the main menu
Adds an animated background to the main menu. 
Also adds pretty buttons and lets the player move them exactly as changing ships in the game. 
Adds visual cues to what keys to press and when they are pressed. 
Adds background music and sound effects for each action in the menu. 
Changes the credits' window to match the game's aesthetics. 
Changes the font to OpenDyslexic font to help people read and it was pretty. But we should add it to be optional later.

Note: Q & E navigate through the menu and Select confirm the option.